components dir restructure
