+++ package.json
@@ -44,13 +44,14 @@
     "husky": "^3.0.9",
     "image-webpack-loader": "^5.0.0",
     "jsdoc": "^3.6.3",
+    "node-notifier": "^8.0.0",
     "prettier": "1.18.2",
     "pretty-quick": "^2.0.0",
     "react-hot-loader": "^4.12.10",
     "style-loader": "^1.2.1",
     "webpack": "^4.39.1",
+    "webpack-bundle-analyzer": "^3.9.0",
     "webpack-cli": "^3.3.6",
-    "node-notifier": "^8.0.0",
     "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
