+++ src/content-scripts/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/App";
+import App from "./App";
 import constants from "../../constants";
 const mountId = constants.contentScript.mountId;
 console.log(mountId);
