+++ src/popup-page/faq.jsx
